server/requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiolimiter==1.1.0
aiopg==1.4.0
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
boto3==1.28.35
botocore==1.31.35
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
croniter==1.4.1
cryptography==41.0.3
dataclasses-json==0.5.14
decorator==5.1.1
executing==1.2.0
frozenlist==1.4.0
google-api-core==2.11.1
google-auth==2.22.0
googleapis-common-protos==1.60.0
h11==0.14.0
html5tagger==1.3.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
ibm-generative-ai==0.2.5
idna==3.4
ipython==8.14.0
jedi==0.19.0
jmespath==1.0.1
joblib==1.3.2
langchain==0.0.276
langsmith==0.0.26
marshmallow==3.20.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
openai==0.27.9
packaging==23.1
pandas==2.0.3
parso==0.8.3
pdfminer.six==20221105
pdfplumber==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
procrastinate==0.27.0
prompt-toolkit==3.0.39
protobuf==4.24.1
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.6.3
Pygments==2.16.1
PyJWT==2.8.0
pypdf==3.15.4
pypdfium2==4.19.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
rsa==4.9
s3transfer==0.6.2
sanic==23.6.0
sanic-ext==23.6.0
sanic-routing==23.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.3.3
tqdm==4.66.1
tracerite==1.1.0
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
urllib3==1.26.16
uvloop==0.17.0
wcwidth==0.2.6
websockets==11.0.3
yarl==1.9.2
